requirements/base.txt
# base requirements

# Django
# --------------------------------------------------------------
Django==3.2.13
django-environ==0.8.1
argon2-cffi==21.3.0  # password
Pillow==9.1.1

# Rest Framework
# --------------------------------------------------------------
django-filter==21.1
django-sql-utils==0.6.1

# Database
# --------------------------------------------------------------
mysqlclient==2.1.0
pymysql==1.0.2